% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBB.R
\name{allelicRatio}
\alias{allelicRatio}
\title{Fit betabinomial on each cell type groups}
\usage{
allelicRatio(sce, level = 0.95, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment containing assays
(\code{"ratio"}, \code{"counts"}) and colData (\code{"x"},
\code{"part"})}

\item{level}{the credible interval required}

\item{...}{Argument for the \code{\link[apeglm]{apeglm}}
functions.}
}
\value{
Posterior mean (\code{"ar"}) for allelic ratio estimator is returned in
rowData for each cell type as well as \code{"s"} value and credible interval(\code{"lower"} and \code{"upper"}).
}
\description{
Fit betabinomial on each cell type groups
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = seq_len(4))
sce_sub <- wilcoxExt(sce, genecluster = 1)
sce_sub <- allelicRatio(sce_sub)
}
