% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusPart.R
\name{consensusPart}
\alias{consensusPart}
\title{Consensus Partitions}
\usage{
consensusPart(sce)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\description{
Derive consensus partitions of an ensemble fused lasso partitions.
}
