% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDisp.R
\name{estDisp}
\alias{estDisp}
\title{Estimate overdispersion parameter of a betabinomial}
\usage{
estDisp(se, ct)
}
\arguments{
\item{se}{SummarizedExperiment with \code{ase} and \code{total}}

\item{ct}{the name of the cell type to be estimated. Default is the highest one.}
}
\description{
Estimate overdispersion parameter of a betabinomial
}
