% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{genecluster}
\alias{genecluster}
\title{Gene clustering based on allelic ratio matrix with pseudo count}
\usage{
genecluster(
  ratio,
  nct,
  G = c(4, 8, 12, 16, 20),
  method = "GMM",
  plot = TRUE,
  ...
)
}
\arguments{
\item{ratio}{an allelic ratio matrix with pseudo count added in}

\item{nct}{number of cell types or states}

\item{method}{either \code{"GMM"} or \code{"hierarchical"}}
}
\description{
cluster on genes
}
