% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimulatedata.R
\name{makeSimulatedata}
\alias{makeSimulatedata}
\title{Make simulate data for airpart}
\usage{
makeSimulatedata(mu1, mu2, nct, n, ngenecl, theta)
}
\arguments{
\item{mu1}{low count}

\item{mu2}{high count}

\item{nct}{number of cell types}

\item{n}{number of cells per cell types}

\item{ngenecl}{number of gene per gene cluster}

\item{theta}{overdispersion parameter}
}
\value{
a list with the following elements:
\item{ase.mat}{maternal allelic expression matrix}
\item{ase.pat}{paternal allelic expression matrix}
\item{x}{a vector of annotated cell types in the same order as cells in count matrix}
}
\description{
Make simulate data for airpart
}
