% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimulatedData.R
\name{makeSimulatedData}
\alias{makeSimulatedData}
\title{Make simulated data for airpart}
\usage{
makeSimulatedData(mu1, mu2, nct, n, ngenecl, theta)
}
\arguments{
\item{mu1}{low count (typical of "noisy" ratio estimates)}

\item{mu2}{high count}

\item{nct}{number of cell types}

\item{n}{number of cells per cell type}

\item{ngenecl}{number of genes per gene cluster}

\item{theta}{overdispersion parameter (higher is closer to binomial)}
}
\value{
a list with the following elements:
\itemize{
  \item{ase.mat}{maternal allelic expression matrix}
  \item{ase.pat}{paternal allelic expression matrix}
  \item{x}{a vector of annotated cell types in the same order as cells in count matrix}
}
}
\description{
Make simulated data for airpart
}
