% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStep.R
\name{makeStep}
\alias{makeStep}
\title{Plot group partition and posterior allelic ratio estimates by step}
\usage{
makeStep(sce, xlab = "cell type")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{xlab}{the x axis name.}
}
\value{
a ggplot2 object.
}
\description{
Plot group partition and posterior allelic ratio estimates by step
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = 1:4)
sce_sub <- wilcoxExt(sce, genecluster = 1)
sce_sub <- allelicRatio(sce_sub)
makeStep(sce_sub)
}
