% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeViolin.R
\name{makeViolin}
\alias{makeViolin}
\title{Plot allelic ratio as heatmap}
\usage{
makeViolin(sce)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\value{
a ggplot2 object
}
\description{
Plot allelic ratio as heatmap
}
\examples{
sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = 1:4)
sce_sub <- wilcoxExt(sce, genecluster = 1)
makeViolin(sce_sub)
}
