% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the SummarizedExperiment}
\usage{
preprocess(se, pc = 2)
}
\arguments{
\item{se}{SummarizedExperiment with \code{ase} and \code{total}}

\item{pc}{pseudocount for calculating the
smoothed ratio}
}
\value{
SummarizedExperiment with total count,
ratio and pseudocount-smoothed ratio, where the
ratio provides the maternal allele count over total
}
\description{
Preprocess the SummarizedExperiment
}
