% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryAllelicRatio}
\alias{summaryAllelicRatio}
\title{Allelic ratio summary}
\usage{
summaryAllelicRatio(se, genecluster)
}
\arguments{
\item{se}{SummarizedExpeirment}

\item{genecluster}{an optional vector of gene cluster IDs.
if nothing is given, all clusters' summary will be displayed}
}
\value{
a list of gene cluster's summary table contains
\itemize{
  \item{weighted.mean} {weighted mean of allelic ratio for the cell types}
  \item{mean} {mean allelic ratio for the cell types}
  \item{var} {variance of allelic ratio for the cell types}
}
}
\description{
produce allelic ratio summary for each gene cluster
}
\examples{
summaryAllelicRatio(se,genecluster = c(1,3))

}
