% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcox}
\alias{wilcox}
\title{Extension on Pairwise Mann Whitney Wilcoxon Test for partitioning}
\usage{
wilcox(data, threshold = 0.05, p.adjust.method = "none", ...)
}
\arguments{
\item{data}{A data frame containing the model response (\code{ratio}),
and grouping factors (\code{x})}

\item{threshold}{a vector with candidate thresholds for raw p-value
cut-off. For details please see vignette}

\item{p.adjust.method}{method for adjusting p-values
(see \code{\link[stats]{p.adjust}}). Can be abbreviated}
}
\value{
A vector grouping factor partition is returned
(TODO: recommend having same output as fusedlass()...x )
}
\description{
Extends the Pairwise Mann Whitney Wilcoxon Test by combining
hierarchical clustering for partition.
}
