% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{allelicRatio}
\alias{allelicRatio}
\title{Fit betabinomial on each cell type groups}
\usage{
allelicRatio(
  se,
  level = 0.95,
  method = c("Normal", "bootstrap"),
  R,
  ncores,
  ...
)
}
\arguments{
\item{se}{A SummarizedExpeirment containing assays (\code{"ratio"},
\code{"total"}) and colData (\code{"x"}, \code{"part"})}

\item{level}{the confidence interval required}

\item{method}{the method to used for construct confidence interval. The default is the first method.
"bootstrap" is likely to be more accurate especially when the data is highly overdispersed,
however it is computationally intensive.}

\item{R}{The number of bootstrap replicates.}

\item{ncores}{integer: number of processes to be used in parallel operation.
Typically one would chose this to the number of available CPUs.}
}
\description{
Fit betabinomial on each cell type groups
}
\examples{
# Use normal approximation to calculate Wald-type confidence intervals
allelicRatio(se)
allelicRatio(se, method = "bootstrap", R = 200, ncores = 4)
}
