% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinom.R
\name{allelicRatio}
\alias{allelicRatio}
\title{Fit betabinomial on each cell type groups}
\usage{
allelicRatio(
  sce,
  level = 0.95,
  method = c("normal", "bootstrap"),
  type = "perc",
  R,
  trace = TRUE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment containing assays
(\code{"ratio"}, \code{"counts"}) and colData (\code{"x"},
\code{"part"})}

\item{level}{the confidence interval required}

\item{method}{the method to used for construct confidence
interval. The default is \code{"normal"}.  \code{"bootstrap"} is
likely to be more accurate especially when the data is highly
overdispersed, however it is computationally intensive.}

\item{type}{the type of intervals required for bootstrap. The values
should be one of the values \code{c("norm","basic","perc")}}

\item{R}{The number of bootstrap replicates.}

\item{trace}{logical indicating if output should be produced for each
iteration.}

\item{...}{Argument for the \code{\link[boot]{boot}}
functions.}
}
\value{
A matrix allelic ratio estimator is returned in metadata
  \code{"estimator"} which includes \code{"estimate"},\code{"std.error"},
  \code{"pvalue"} indicated allelic imbalance significance
  which calculated by \code{t} statistics
  and \code{"confidence interval"}
}
\description{
Fit betabinomial on each cell type groups
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = seq_len(4))
sce_sub <- wilcoxExt(sce, genecluster = 1)

# Use normal approximation to calculate
# Wald-type confidence intervals
sce_sub <- allelicRatio(sce_sub)

# Alternative with bootstrap
sce_sub <- allelicRatio(sce_sub,
  method = "bootstrap", R = 5, type="norm",
  parallel = "multicore", ncpus = 4
)

}
