% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{allelicRatio}
\alias{allelicRatio}
\title{Fit betabinomial on each cell type groups}
\usage{
allelicRatio(sce, level = 0.95, method = c("Normal", "bootstrap"), R, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment containing assays (\code{"ratio"},
\code{"counts"}) and colData (\code{"x"}, \code{"part"})}

\item{level}{the confidence interval required}

\item{method}{the method to used for construct confidence interval. The default is the first method.
"bootstrap" is likely to be more accurate especially when the data is highly overdispersed,
however it is computationally intensive.}

\item{R}{The number of bootstrap replicates.}

\item{...}{Argument for the \code{\link[boot]{boot}} functions.}
}
\value{
A matrix allelic ratio estimator is returned in
metadata \code{"estimator"}
}
\description{
Fit betabinomial on each cell type groups
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G=1:4)
sce_sub <- wilcoxExt(sce,genecluster=1)

# Use normal approximation to calculate Wald-type confidence intervals
sce_sub <- allelicRatio(sce_sub)

# Alternative with bootstrap
\dontrun{
sce_sub <- allelicRatio(sce_sub, method = "bootstrap", R = 200,
parallel="multicore", ncpus = 4)
}

}
