% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{betaBinom}
\alias{betaBinom}
\title{Fit betabinomial on each cell type groups}
\usage{
betaBinom(se, ...)
}
\arguments{
\item{se}{A SummarizedExpeirment containing assays (\code{"ratio"},
\code{"total"}) and colData (\code{"x"}, \code{"part"})}
}
\description{
Fit betabinomial on each cell type groups
}
