% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{betabinom}
\alias{betabinom}
\title{Fit betabinomial on each cell type groups}
\usage{
betabinom(data, part, ...)
}
\arguments{
\item{data}{A data frame containing:
the allelic ratio (\code{"ratio"}),
and the total counts for weighting (\code{"cts"})
the cell type group (\code{"part"})}

\item{part}{the column name of cell type group}
}
\value{
a list with the following elements:
\item{coef_bb}{the allelic ratio estimator for each cell type group}
\item{confint_wilcoxon}{the allelic ratio confindence interval for each cell type group}
\item{rho}{the correlation between the N individuals within a group. rho is given by 1/(1 + alpha + beta) and is
It is known as the over-dispersion parameter}
}
\description{
Fit betabinomial on each cell type groups
}
