% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusPart.R
\name{consensusPart}
\alias{consensusPart}
\title{Consensus Partitions}
\usage{
consensusPart(se)
}
\arguments{
\item{se}{SummarizedExpeirment}
}
\description{
Derive consensus partitions of an ensemble fused lasso partitions.
}
