% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDisp.R
\name{estDisp}
\alias{estDisp}
\title{Estimate overdispersion parameter of a betabinomial}
\usage{
estDisp(sce, pc = 2, genecluster, type = c("plot", "values"))
}
\arguments{
\item{sce}{SingleCellExperiment with \code{a1} matrix and \code{counts}}

\item{pc}{pseudocount for calculating the smoothed
ratio in the preprocess step.}

\item{genecluster}{which gene cluster dispersion
parameter want to be estimated.
Default is the cluster with the most cells}

\item{type}{indicated whether to output
dispersion estimates as a plot or a value}
}
\value{
A ggplot object of the dispersion estimates over the mean
}
\description{
Estimate overdispersion parameter of a betabinomial
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = seq_len(4))
estDisp(sce)

}
