% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedlasso.R
\name{fusedlasso}
\alias{fusedlasso}
\title{Generalized fused lasso to partition cell types by allelic imbalance}
\usage{
fusedlasso(
  formula,
  model = "binomial",
  data,
  niter = 1,
  pen.weights,
  lambda = "cv1se.dev",
  k = 5,
  adj.matrix,
  lambda.length = 25L,
  se.rule.nct = 8,
  se.rule.mult = 0.5,
  ...
)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} object which will typically be:
\code{ratio ~ p(x, pen="gflasso")}.
See \code{\link[smurf]{glmsmurf}} for more details}

\item{model}{Either \code{"binomial"} or \code{"gaussian"} used to fit the generalized fused lasso}

\item{data}{A data frame containing:
the allelic ratio (\code{"ratio"}),
the cell assignment (\code{"x"}),
and the total counts for weighting (\code{"cts"})}

\item{niter}{number of iteration to run; recommended to run 5 times
if allelic ratio differences are within [0.05,0.1]}

\item{pen.weights}{argument as described in \code{\link[smurf]{glmsmurf}}}

\item{lambda}{argument as described in \code{\link[smurf]{glmsmurf}}}

\item{k}{number of cross-validation folds}

\item{adj.matrix}{argument as described in \code{\link[smurf]{glmsmurf}}}

\item{lambda.length}{argument as described in \code{\link[smurf]{glmsmurf}}}

\item{se.rule.nct}{the number of cell types to trigger an SE based rule
(to prioritize smaller models, more fusing). When the number of cell types
is less than or equal to this value, the SE rule is used}

\item{se.rule.mult}{the multiplier of the SE in determining the lambda:
the chosen lambda is within \code{se.rule.mult} x SE of the minimum
deviance}

\item{...}{additional arguments passed to \code{\link[smurf]{glmsmurf}}}
}
\value{
An object of class 'glmsmurf' is returned.
}
\description{
Fits generalized fused lasso with either binomial(link="logit")
or gaussian likelihood, leveraging functions from the
\code{smurf} package.
}
\details{
See the package vignette for more details and a complete description of a use case.
}
\references{
This function leverages the glmsmurf function from the smurf package.
For more details see the following manuscript:

Sander Devriendt, Katrien Antonio, Tom Reynkens, Roel Verbelen
"Sparse Regression with Multi-type Regularized Feature Modeling"
2018. arXiv:1810.03136 [stat.CO].
}
\seealso{
\code{\link[smurf]{glmsmurf}}, \code{\link[smurf]{glmsmurf.control}},
\code{\link[smurf]{p}}, \code{\link[stats]{glm}}
}
