% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{geneCluster}
\alias{geneCluster}
\title{Gene clustering based on allelic ratio matrix with pseudo-count}
\usage{
geneCluster(se, G = c(8, 12, 16, 20, 24), method = "GMM", plot = TRUE, ...)
}
\arguments{
\item{se}{SummarizedExpeirment containing assays \code{"ratio_pseudo"} and
colData \code{"x"}}

\item{method}{either \code{"GMM"} or \code{"hierarchical"}}

\item{plot}{logical, whether to make a PCA plot}
}
\value{
gene cluster IDs are stored in the rowData column \code{cluster}
}
\description{
Gene clustering based on allelic ratio matrix with pseudo-count
}
\references{
This function leverages Mclust from the mclust package, or hclust.

For mclust see:
Luca Scrucca and Michael Fop and T. Brendan Murphy, Adrian E. Raftery
"mclust 5: clustering, classification and density
estimation using {G}aussian finite mixture models"
2016. The R Journal. doi: 10.32614/RJ-2016-021
}
\seealso{
\code{\link[mclust]{Mclust}}
}
