% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeForest.R
\name{makeForest}
\alias{makeForest}
\title{Extension on Pairwise Mann Whitney Wilcoxon Test for partitioning}
\usage{
makeForest(
  ar,
  xticks,
  xlab = "Allelic Ratio",
  col = fpColors(),
  grid = structure(c(0.1, 0.5, 0.9), gp = gpar(lty = 2, col = "#CCCCFF")),
  ...
)
}
\arguments{
\item{ar}{A data frame containing: the allelic ratio estimator ("estimator"),
and last two column is the confidence interval.}

\item{xticks}{argument as described in \code{\link[forestplot]{forestplot}}}

\item{xlab}{x-axis label. Default is "Allelic Ratio"}

\item{col}{Set the colors for all the elements. See \code{\link[forestplot]{fpColors}} for details}

\item{grid}{If you want a discrete gray dashed grid at the level of the ticks
you can set this parameter to TRUE.
If you set the parameter to a vector of values lines will be drawn at the corresponding positions.
If you want to specify the \code{\link[grid]{gpar}} of the lines then either directly pass a \code{\link[grid]{gpar}} object
or set the gp attribute e.g. attr(line_vector, "gp") <- \code{\link[grid]{gpar}}(lty=2, col = "red")}

\item{boxsize}{Override the default box size based on precision}
}
\description{
Draw a forest plot to visualized cell type specific allelic ratio estimator. It is based on the \pkg{rmeta}-package`s
\code{forestplot} function.
}
\examples{
xticks=seq(from = 0, to = 1, by = 0.05)
xtlab <- rep(c(TRUE, FALSE), length.out = length(xticks))
attr(xticks, "labels") <- xtlab
makeForest(ar,xticks,col=fpColors(box = "royalblue"))

}
\seealso{
\code{\link[forestplot]{forestplot}}, \code{\link[forestplot]{fpColors}},
\code{\link[forestplot]{fpShapesGp}}, \code{\link[forestplot]{fpLegend}}
}
