% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeRatioHeatmap}
\alias{makeRatioHeatmap}
\title{Plot allelic ratio as heatmap}
\usage{
makeRatioHeatmap(sce, show_row_names = FALSE, ...)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{show_row_names}{show row names or not}

\item{...}{Passsed on the other argument in \code{\link[ComplexHeatmap]{Heatmap}}.}
}
\description{
Plot allelic ratio as heatmap
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
makeRatioHeatmap(sce) # display whole data allelic ratio pattern

sce <- geneCluster(sce, G=1:4)
sce_sub <- wilcoxExt(sce,genecluster=1)
sce_sub <- allelicRatio(sce_sub)
makeRatioHeatmap(sce_sub) # display specific gene cluster partition result

}
