% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeRatioHeatmap}
\alias{makeRatioHeatmap}
\title{Plot allelic ratio as heatmap}
\usage{
makeRatioHeatmap(sce, show_row_names = FALSE, ...)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{show_row_names}{show row names or not}

\item{...}{Passsed on the other argument in \code{\link[ComplexHeatmap]{Heatmap}}.}
}
\description{
Plot allelic ratio as heatmap
}
