% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeRatioHeatmap}
\alias{makeRatioHeatmap}
\title{Plot allelic ratio as heatmap}
\usage{
makeRatioHeatmap(se, show_row_names = FALSE, ...)
}
\arguments{
\item{se}{SummarizedExpeirment}

\item{show_row_names}{show row names or not}
}
\description{
Plot allelic ratio as heatmap
}
