% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeRatioHeatmap}
\alias{makeRatioHeatmap}
\title{Plot allelic ratio as heatmap}
\usage{
makeRatioHeatmap(se)
}
\arguments{
\item{se}{SummarizedExpeirment}
}
\description{
Plot allelic ratio as heatmap
}
