% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeViolin.R
\name{makeViolin}
\alias{makeViolin}
\title{Smoothed allelic ratio in violin plots}
\usage{
makeViolin(sce)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\value{
a ggplot2 object
}
\description{
Smoothed allelic ratio in violin plots
}
\examples{

sce <- makeSimulatedData()
sce <- preprocess(sce)
sce <- geneCluster(sce, G = 1:4)
sce_sub <- wilcoxExt(sce, genecluster = 1)
makeViolin(sce_sub)

}
