% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the SingleCellExperiment}
\usage{
preprocess(sce, pc = 2)
}
\arguments{
\item{sce}{SingleCellExperiment with \code{ase.mat} and \code{ase.pat}}

\item{pc}{pseudocount for calculating the
smoothed ratio}
}
\value{
SingleCellExperiment with total count,
ratio and pseudocount-smoothed ratio, where the
ratio provides the maternal allele count over total
}
\description{
Preprocess the SingleCellExperiment
}
