% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcox}
\alias{wilcox}
\title{Extension on Pairwise Mann Whitney Wilcoxon Test for partition}
\usage{
wilcox(data, threshold = 0.05, p.adjust.method = "none", ...)
}
\arguments{
\item{data}{A data frame containing the model response (\code{ratio}), and grouping factors (\code{x})}

\item{threshold}{a vector with candidate thresholds as p-value cut-off. Details please look at vignette}

\item{p.adjust.method}{method for adjusting p values (see \code{\link[stats]{p.adjust}}). Can be abbreviated}
}
\value{
A vector of grouping factor partition is returned.
}
\description{
Extend the Pairwise Mann Whitney Wilcoxon Test by combining hierarchical cluster for partition.
}
