% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxExt}
\alias{wilcoxExt}
\title{Extension on Pairwise Mann Whitney Wilcoxon Test for partitioning}
\usage{
wilcoxExt(sce, genecluster, threshold, p.adjust.method = "none", ...)
}
\arguments{
\item{sce}{A SingleCellExperiment containing assays (\code{"ratio"},
\code{"counts"}) and colData \code{"x"}}

\item{genecluster}{which gene cluster result want to be returned.
Usually identified interesting gene cluster pattern by \code{\link{summaryAllelicRatio}}}

\item{threshold}{a vector with candidate thresholds for raw p-value
cut-off. For details please see vignette}

\item{p.adjust.method}{method for adjusting p-values
(see \code{\link[stats]{p.adjust}}). Can be abbreviated}

\item{...}{additional arguments to pass to \code{\link[stats]{wilcox.test}}.}
}
\value{
A vector grouping factor partition is returned
}
\description{
Extends the Pairwise Mann Whitney Wilcoxon Test by combining
hierarchical clustering for partition.
}
